Tester: added integer question infrastructure.  KP-5153

+++ src/Core/SharedKernels/DataCollection/DataCollection/Implementation/Entities/InterviewQuestion.cs
@@ -99,7 +99,7 @@ namespace WB.Core.SharedKernels.DataCollection.Implementation.Entities
 
     public class IntegerNumericAnswer : BaseInterviewAnswer
     {
-        public long Answer { get; set; }
+        public int Answer { get; set; }
 
         public IntegerNumericAnswer() { }
         public IntegerNumericAnswer(Guid id, decimal[] rosterVector)
